+++ website/src/pages/PRsPage/PRsPage.tsx
@@ -21,7 +21,7 @@ import RingLoader from "react-spinners/RingLoader";
 
 import PRTable from "./components/PRTable";
 import { prDataTypes } from "@/types";
-import Hero, { HeroProps } from "@/common/Hero";
+import PRHero from "./components/PRHero";
 
 export default function PRsPage() {
   const {
@@ -30,23 +30,9 @@ export default function PRsPage() {
     error: PRListError,
   } = useApiCall<prDataTypes>(`${import.meta.env.VITE_API_URL}pr/list`);
 
-  const heroProps: HeroProps = {
-    title: "Pull Request",
-    description: (
-      <p>
-      If a given Pull Request on vitessio/vitess is labelled with the
-      <span className="bg-red-800 text-white px-2 py-1 rounded-2xl ml-2">Benchmark Me</span> label
-      the Pull Request will be handled and benchmarked by arewefastyet. For each commit on the Pull Request there will be two benchmarks: one on the Pull Request's HEAD and another on the base of the Pull Request.
-      <br />
-      <br />
-      On this page you can find all benchmarked Pull Requests.
-    </p>
-    ),
-  };
-
   return (
     <>
-      <Hero title={heroProps.title} description={heroProps.description}/>
+      <PRHero/>
 
       {isPRListLoading && (
         <div className="flex justify-center w-full my-16">
